separate the widgets out into individual files
